Fix version information injection via tarball
