+++ packages/bundle/src/boot/actual/full-es5.ts
@@ -1,8 +1,8 @@
-import { buildInfo as minimalBuildInfo, version } from './exports-minimal';
-import defaultCreateDirectLine from '../createDirectLine';
-import defaultCreateDirectLineAppServiceExtension from '../createDirectLineAppServiceExtension';
+import { buildInfo as minimalBuildInfo, version } from './minimal';
+import defaultCreateDirectLine from '../../createDirectLine';
+import defaultCreateDirectLineAppServiceExtension from '../../createDirectLineAppServiceExtension';
 
-export * from './exports';
+export * from './full';
 
 export const buildInfo = Object.freeze({ ...minimalBuildInfo, variant: 'full-es5' });
