+++ packages/bundle/src/boot/addVersion.ts
