+++ packages/bundle/src/boot/exports/full-es5.ts
@@ -1,0 +1,8 @@
+import { buildInfo as actualBuildInfo } from '../actual/full-es5';
+import addVersion from '../addVersion';
+
+export * from '../actual/full-es5';
+
+export const buildInfo = Object.freeze({ ...actualBuildInfo, variant: 'full-es5' });
+
+addVersion(buildInfo);
